TS definitions for assignments and absorb plan.
