+++ crates/but-hunk-assignment/src/lib.rs
@@ -24,15 +24,21 @@ use itertools::Itertools;
 use reconcile::MultipleOverlapping;
 use serde::{Deserialize, Serialize};
 use tracing::instrument;
+use ts_rs::TS;
 use uuid::Uuid;
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, TS)]
 #[serde(rename_all = "camelCase")]
+#[cfg_attr(
+    feature = "export-ts",
+    ts(export, export_to = "./hunkAssignment/index.ts")
+)]
 pub struct HunkAssignment {
     /// A stable identifier for the hunk assignment.
     ///   - When a new hunk is first observed (from the uncommitted changes), it is assigned a new id.
     ///   - If a hunk is modified (i.e. it has gained or lost lines), the UUID remains the same.
     ///   - If two or more hunks become merged (due to edits causing the contexts to overlap), the id of the hunk with the most lines is adopted.
+    #[ts(type = "string | null")]
     pub id: Option<Uuid>,
     /// The hunk that is being assigned. Together with path_bytes, this identifies the hunk.
     /// If the file is binary, or too large to load, this will be None and in this case the path name is the only identity.
@@ -40,8 +46,10 @@ pub struct HunkAssignment {
     /// The file path of the hunk.
     pub path: String,
     /// The file path of the hunk in bytes.
+    #[ts(type = "number[]")]
     pub path_bytes: BString,
     /// The stack to which the hunk is assigned. If None, the hunk is not assigned to any stack.
+    #[ts(type = "string | null")]
     pub stack_id: Option<StackId>,
     /// The dependencies(locks) that this hunk has. This determines where the hunk can be assigned.
     /// This field is ignored when HunkAssignment is passed by the UI to create a new assignment.
@@ -126,6 +134,103 @@ impl From<HunkAssignment> for but_core::DiffSpec {
             hunk_headers,
         }
     }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default, TS)]
+#[serde(rename_all = "camelCase", tag = "type", content = "subject")]
+#[cfg_attr(
+    feature = "export-ts",
+    ts(export, export_to = "./hunkAssignment/index.ts")
+)]
+pub enum AbsorptionTarget {
+    Branch {
+        branch_name: String,
+    },
+    HunkAssignments {
+        assignments: Vec<HunkAssignment>,
+    },
+    #[default]
+    All,
+}
+
+/// Reason why a file is being absorbed to a particular commit
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, TS)]
+#[serde(rename_all = "snake_case")]
+#[cfg_attr(
+    feature = "export-ts",
+    ts(export, export_to = "./hunkAssignment/index.ts")
+)]
+pub enum AbsorptionReason {
+    /// File has hunk range overlap with this commit
+    HunkDependency,
+    /// File is assigned to this stack and this is the topmost commit
+    StackAssignment,
+    /// Default to leftmost stack's topmost commit
+    DefaultStack,
+}
+
+impl AbsorptionReason {
+    pub fn description(&self) -> &str {
+        match self {
+            AbsorptionReason::HunkDependency => "files locked to commit due to hunk range overlap",
+            AbsorptionReason::StackAssignment => "last commit in the assigned stack",
+            AbsorptionReason::DefaultStack => "last commit in the primary lane",
+        }
+    }
+}
+
+/// Information about a file being absorbed
+#[derive(Debug, Clone, Serialize, Deserialize, TS)]
+#[serde(rename_all = "camelCase")]
+#[cfg_attr(
+    feature = "export-ts",
+    ts(export, export_to = "./hunkAssignment/index.ts")
+)]
+pub struct FileAbsorption {
+    pub path: String,
+    pub assignment: HunkAssignment,
+}
+
+/// Information about absorptions grouped by commit
+#[derive(Debug, Serialize, Deserialize, TS)]
+#[serde(rename_all = "camelCase")]
+#[cfg_attr(
+    feature = "export-ts",
+    ts(export, export_to = "./hunkAssignment/index.ts")
+)]
+pub struct CommitAbsorption {
+    #[ts(type = "string")]
+    pub stack_id: but_core::ref_metadata::StackId,
+    #[ts(type = "string")]
+    #[serde(with = "but_serde::object_id")]
+    pub commit_id: gix::ObjectId,
+    pub commit_summary: String,
+    pub files: Vec<FileAbsorption>,
+    pub reason: AbsorptionReason,
+}
+
+/// JSON output structure for a file being absorbed
+#[derive(Debug, Serialize)]
+pub struct JsonFileAbsorption {
+    pub path: String,
+    pub hunks: Vec<String>,
+}
+
+/// JSON output structure for a commit absorption
+#[derive(Debug, Serialize)]
+pub struct JsonCommitAbsorption {
+    pub commit_id: String,
+    pub commit_summary: String,
+    pub reason: AbsorptionReason,
+    pub reason_description: String,
+    pub files: Vec<JsonFileAbsorption>,
+}
+
+/// JSON output structure for the entire absorb operation
+#[derive(Debug, Serialize)]
+pub struct JsonAbsorbOutput {
+    pub total_files: usize,
+    pub commits: Vec<JsonCommitAbsorption>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
